refactor ♻️: Enhance CustomResponse class with new traits for improved API response handling
- Implemented CommonApiResponsesTrait and CustomResponseMethodsTrait to streamline response methods and enhance code reusability.
- Updated CustomResponse class to implement Responsable interface, allowing for more flexible response handling.
- Removed redundant methods from CustomResponse and shifted functionality to the new traits, improving code organization and maintainability.
